[rpcli] properties.cpp, ListDataField: Changed `totalWidth` from `unsigned int` to `size_t`.
This fixes the following warning with MSVC 2017 v15.9.15:

src\rpcli\properties.cpp(319): warning C4267: '+=': conversion from 'size_t' to 'unsigned int', possible loss of data

+++ src/rpcli/properties.cpp
@@ -252,7 +252,7 @@ public:
 
 		// Column names
 		unique_ptr<size_t[]> colSize(new size_t[col_count]());
-		unsigned int totalWidth = col_count + 1;
+		size_t totalWidth = col_count + 1;
 		if (listDataDesc.names) {
 			int i = 0;
 			for (auto it = listDataDesc.names->cbegin(); it != listDataDesc.names->cend(); ++it, ++i) {
